05-Renato Fernandes/Analista_juridico-1/requirements.txt
streamlit>=1.35.0
openai>=1.30.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.20.0

nltk>=3.9.0
reportlab>=4.0.0
